+++ central/probeupload/manager/manager.go
@@ -13,4 +13,8 @@ type Manager interface {
 	GetExistingProbeFiles(ctx context.Context, files []string) ([]*v1.ProbeUploadManifest_File, error)
 
 	StoreFile(ctx context.Context, file string, data io.Reader, size int64, crc32 uint32) error
+
+	// OpenFile attempts to open a probe file, returning the data reader and its size, or an error. This function does
+	// not perform any access checks.
+	LoadProbe(ctx context.Context, file string) (io.ReadCloser, int64, error)
 }