+++ central/probeupload/service/service_impl.go
@@ -9,15 +9,21 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/pkg/errors"
+	"github.com/stackrox/rox/central/clusters"
 	"github.com/stackrox/rox/central/probeupload/manager"
 	"github.com/stackrox/rox/central/role/resources"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/pkg/auth/permissions"
+	"github.com/stackrox/rox/pkg/env"
 	"github.com/stackrox/rox/pkg/grpc/authz"
+	"github.com/stackrox/rox/pkg/grpc/authz/idcheck"
 	"github.com/stackrox/rox/pkg/grpc/authz/perrpc"
 	"github.com/stackrox/rox/pkg/grpc/authz/user"
 	"github.com/stackrox/rox/pkg/grpc/routes"
 	"github.com/stackrox/rox/pkg/httputil"
+	"github.com/stackrox/rox/pkg/httputil/proxy"
+	"github.com/stackrox/rox/pkg/kocache"
+	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/probeupload"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -29,15 +35,31 @@ var (
 			"/v1.ProbeUploadService/GetExistingProbes",
 		},
 	})
+
+	log = logging.LoggerForModule()
 )
 
 type service struct {
 	mgr manager.Manager
+
+	probeServerHandler http.Handler
 }
 
 func newService(mgr manager.Manager) *service {
+	var probeSources []probeupload.ProbeSource
+	probeSources = append(probeSources, mgr)
+	if env.OfflineModeEnv.Setting() != "true" {
+		if baseURL := clusters.CollectorModuleDownloadBaseURL.Setting(); baseURL != "" {
+			httpClient := &http.Client{
+				Transport: proxy.RoundTripper(),
+			}
+			probeSources = append(probeSources, kocache.New(context.Background(), httpClient, baseURL, kocache.Options{}))
+		}
+	}
+
 	return &service{
-		mgr: mgr,
+		mgr:                mgr,
+		probeServerHandler: probeupload.NewProbeServerHandler(probeupload.LogCallback(log), probeSources...),
 	}
 }
 
@@ -69,6 +91,12 @@ func (s *service) CustomRoutes() []routes.CustomRoute {
 			Route:         "/api/extensions/probeupload",
 			Authorizer:    user.With(permissions.Modify(resources.ProbeUpload)),
 			ServerHandler: http.HandlerFunc(s.handleProbeUpload),
+			Compression:   false,
+		},
+		{
+			Route:         "/kernel-objects/",
+			Authorizer:    idcheck.SensorsOnly(),
+			ServerHandler: http.StripPrefix("/kernel-objects", s.probeServerHandler),
 			Compression:   false,
 		},
 	}